+++ CHANGELOG.md
@@ -42,7 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - improved error msg's for failures
 
 ### Fixed
-- pci speed/width in the snapshot
+- PCI speed/width in the snapshot
 - list option now matches dev/tenstorrent device IDS
 - reset_config generation is on local devices only