+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.0.2 - 09/10/2024
+
+- Bug fix for PCIE Gen1 reporting
+- Added spelling fixes
+
 ## 3.0.1 - 16/09/2024
 
 - Bumped luwen lib versions (0.4.0 -> 0.4.3)